Add lifo cache to new module; drop "utils", bleh
