+++ piker/_cacheables.py
@@ -18,6 +18,7 @@
 Cacheing apis and toolz.
 
 """
+from collections import OrderedDict
 from typing import Optional
 from contextlib import (
     asynccontextmanager,
@@ -34,6 +35,35 @@ from .data.feed import Feed
 
 
 log = get_logger(__name__)
+
+
+def async_lifo_cache(maxsize=128):
+    """Async ``cache`` with a LIFO policy.
+
+    Implemented my own since no one else seems to have
+    a standard. I'll wait for the smarter people to come
+    up with one, but until then...
+    """
+    cache = OrderedDict()
+
+    def decorator(fn):
+
+        async def wrapper(*args):
+            key = args
+            try:
+                return cache[key]
+            except KeyError:
+                if len(cache) >= maxsize:
+                    # discard last added new entry
+                    cache.popitem()
+
+                # do it
+                cache[key] = await fn(*args)
+                return cache[key]
+
+        return wrapper
+
+    return decorator
 
 
 _cache: dict[str, 'Client'] = {}  # noqa
