+++ piker/brokers/questrade.py
@@ -42,11 +42,10 @@ import wrapt
 import asks
 
 from ..calc import humanize, percent_change
-from .._cacheables import open_cached_client
+from .._cacheables import open_cached_client, async_lifo_cache
 from . import config
 from ._util import resproc, BrokerError, SymbolNotFound
 from ..log import get_logger, colorize_json, get_console_log
-from .._async_utils import async_lifo_cache
 from . import get_brokermod
 from . import api
